+++ packages/credential-sdk/src/utils/types/set.js
@@ -1,4 +1,25 @@
-import { ensureSet } from './ensure-type';
+/**
+ * Return true if the given value is a Set.
+ * @param value
+ * @returns {boolean}
+ */
+export function isSet(value) {
+  return value instanceof Set;
+}
+
+/**
+ * Ensures that the given value is a Set. Throws an error with an optional context message if not.
+ * @param {any} value - The value to check.
+ * @param {string} [message] - Optional message for the error.
+ * @returns {Set} - The value if it's a Set.
+ */
+export function ensureSet(value) {
+  if (isSet(value)) {
+    return value;
+  }
+
+  throw new TypeError(`\`${value}\` needs to be a Set.`);
+}
 
 /**
  * Ensures that there is no intersection between two sets