+++ packages/credential-sdk/src/vc/status-list2021-credential.js
@@ -5,7 +5,7 @@ import {
   StatusList, // eslint-disable-line
 } from '@digitalcredentials/vc-status-list';
 import { gzip, ungzip } from 'pako';
-import { u8aToU8a } from '../utils/bytes';
+import { u8aToU8a } from '../utils/types/bytes';
 import VerifiableCredential from './verifiable-credential';
 import { ensureStatusListId } from '../utils';
 import { KeyDoc } from "./helpers"; // eslint-disable-line