+++ scripts/bench/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @docknetwork/benchmarks
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @docknetwork/cheqd-blockchain-modules@0.33.0
+  - @docknetwork/credential-sdk@0.50.0
+  - @docknetwork/cheqd-blockchain-api@0.36.0
+  - @docknetwork/dock-blockchain-modules@0.30.0
+
 ## 0.1.2
 
 ### Patch Changes